requirements.txt
# Core scientific stack
pandas==2.2.2
numpy==1.26.4
openpyxl==3.1.5

# Sentence embeddings
sentence-transformers==2.7.0
faiss-cpu==1.7.4

# Google Gemini API
google-generativeai==0.7.2
python-dotenv==1.0.1

# Data I/O and parquet
pyarrow==15.0.2
fastparquet==2024.2.0

# Fuzzy matching utilities
rapidfuzz==3.9.3

# Optional (for debugging & visualization later)
matplotlib==3.9.2
seaborn==0.13.2
